requirements.txt
torch>=2.0
torchvision>=0.15
numpy>=1.24
scipy>=1.10
scikit-learn>=1.2
ripser>=0.6
persim>=0.3
umap-learn>=0.5
matplotlib>=3.7
